requirements.txt
llama-index-core == 0.11.8
llama-index-readers-file == 0.2.1
llama-index-llms-ollama == 0.3.1
llama-index-embeddings-huggingface == 0.3.1
llama-index-vector-stores-chroma == 0.2.0
fastapi>=0.68.0
